foobar-kafka/owm-producer/requirements.txt
kafka-python
dataprep
